// src/app/model/educacion.model.ts
export class Educacion{
    id: number;
    titulo: String;
    institucion: String;
    descripcion: String;
    fecha_inicio: String;
    fecha_finalizacion: String;
    img: String;

/*     constructor(titulo: String, institucion: String, descripcion: String, fecha_inicio: String, fecha_finalizacion: String, imagen: String){
        this.titulo = titulo;
        this.institucion = institucion;
        this.descripcion = descripcion;
        this.fecha_inicio = fecha_inicio;
        this.fecha_finalizacion = fecha_finalizacion;
        this.imagen = imagen;
    } */
}